feat: Improve style (mainly color setup)
This commit makes it possible to configure the primary color hue via an
environment variable named `CME_PRIMARY_COLOR_HUE`. The primary color
hue is used to generate the primary color palette. The primary color hue
defaults to 231 which is a purple hue that matches the base for the
primary color used in the Capella Collaboration Manager.

This commit also ensures that gray shades are only expressed using so
called `neutral` colors instead of shaded from the `gray` and `slate`
palette.

The vertical alignment of badges on the report index page has been
fixed. The badges are now aligned at the bottom of the card.

A small fix is also applied to the spinner icon to ensure that it has
no dent.

Co-authored-by: Martin Lehmann <martin.lehmann@deutschebahn.com>