perf: Cache rendered reports for Git models
In the case of a model maintained in a Git repository, a rendered report
will be loaded from the cache as long it is not more than a year old or
there is a change for any out of the following:

- the model element UUID (which is part of the URL)
- the template file
- the version of the capella-model-explorer
- the version of the capellambse library
- the version of the capellambse-context-diagrams library
- the model revision

Note, that non-Git model versions cannot be distinguished reliably in
the cache. Therefore, rendered reports for non-Git are not cached at
all.