feat(cli): Always rebuild the CSS before running
When running from a source tree in non-dev mode, the built CSS may still
become stale. With this commit, the CSS will always be rebuilt before
launching, unless the `--skip-build` flag was passed. Note that even
with `--skip-build`, the CSS will still be built initially if it doesn't
exist at all yet.