refactor: Simplify the page layout and route handlers
Use the same container structure for both the main and template pages,
and generally prefer innerHTML swaps wherever it makes sense.

This fixes some minor layouting and styling issues / mismatches. As an
example, the template page placeholder had a border around it, which was
removed once the first template was loaded. By avoiding to swap out the
entire navbar, and instead only replacing the breadcrumbs component
inside it, this also fixes the issue that the dark mode toggle button
would break after navigating to a different page.

Resolves #148