feat: Add an option to disable logfmt
Logfmt logging is great when the logs are being processed automatically.
However, it makes reading the logs for humans a bit more difficult. This
is especially true for multi-line log messages like exception stack
traces, which makes debugging locally more cumbersome than it needs to.

Add a simple log config switch for disabling logfmt and using a more
human-friendly format.