refactor: Replace plot.ly with bigger-picture
Replace a multi-megabyte multi-purpose JS library with a tiny one that
focuses on the one feature we actually wanted in the first place.

This also introduces parcel as a Javascript bundler, as bigger-picture
does not provide easily accessible pre-built public bundles.

Javascript and CSS that is to be bundled was moved from `static` to the
new `frontend` directory. This means that `static` no longer contains
sources and intermediate artifacts that shouldn't be shipped. This
change primarily affects the final Docker image.

Resolves #129