perf: Allow caching reports at least until server shutdown
For models that aren't under strict version control (e.g. local files
without a backing Git repository), caching used to be completely
disabled. This commit enables it for the duration that the server is
alive. While the effect is significantly smaller than the full caching
that Git-backed models allow, it does at least provide some speedups
with virtually no additional maintenance overhead on the server side.