ci: Enable calculating the version number in Docker builds
Since #167, the model explorer version number is shown on the main page.
However, for CI-built docker images, this always shows "v0.1.dev1",
because the Git history is not available by default.

This commit instructs the CI to make a tree-less partial clone instead
of a shallow clone for the Docker build job, which allows setuptools-scm
to calculate the correct version number.